+++ front/src/pages/Make/Snowball/MainDeco/MakeButton.tsx
@@ -1,8 +1,7 @@
 import { useState, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
-import axios from '@utils/axios';
 import styled from 'styled-components';
-import { LongButton } from '@utils';
+import { LongButton, axios } from '@utils';
 import { DecoContext } from './DecoProvider';
 
 interface MakeButtonProps {
