+++ front/src/pages/Visit/Deco/PostButton.tsx
@@ -1,8 +1,7 @@
 import { useContext, useState, useRef } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import axios from '@utils/axios';
 import styled from 'styled-components';
-import { LongButton } from '@utils';
+import { LongButton, axios } from '@utils';
 import { DecoContext } from './DecoProvider';
 import { SnowBallContext } from '../SnowBallProvider';
