+++ front/src/utils/index.tsx
@@ -1,4 +1,5 @@
 export { default as theme } from './theme';
 export { CanvasContainer, Container, LongButton, BlurBody } from './styled';
 export { default as Loading } from './Loading';
+export { default as axios } from './axios';
 export { default as getDecoPoisition } from './position';
